add newsletters switch group and migrate switches over
